CiCo/requirements.txt
pympi-ling
intervaltree
zsvision
mergedeep
beartype
wget
pandas
humanize
mock
tqdm
tensorboard
scipy
pillow
scikit-learn
youtube-dl
frozendict
matplotlib
numpy